requirements.txt
timm==0.4.12
einops
torchprofile
fvcore
